backend/requirements.txt
django-extensions==3.1.5
Django==3.2.16
djangorestframework==3.14.0
djoser==2.1.0
python-decouple==3.5
drf-extra-fields==3.2.1
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.3
django-cors-headers==3.13.0
webcolors==1.12
django-filter==22.1
